First Form Completed (for now) for the habit app
I created a form for creating good habits in the habit app. I will be adding more to the form dynamically with javascript and potentially more through html in the future, but for now I am done with this form. My next form will be for adding in bad habits to slowly quit.