Create Add Bad Habit Form Done (for now)
The add bad habit form is complete for now! I am likely going to add more to both the good habit and bad habit forms in the future but for now I am ready to focus on styling, general html content, and likely tomorrow form validation for the form itself.